feat: Added Question2 resolution (Periodic Keys), mmc function to euclides and modifiend compilators

+++ include/pollard.h
@@ -1,6 +1,6 @@
 #ifndef POLLARD_H
 #define POLLARD_H
 
-long long pollard(int n);
+long long pollard(long long n);
 
 #endif
